Re-Factor Code

+++ src/main/java/com/project/countries/PopulationController.java
@@ -12,26 +12,24 @@ import java.util.ArrayList;
 @RestController
 @RequestMapping("/population")
 public class PopulationController {
+    // localhost:2019/population/size/{people}
     @GetMapping(value = "/size/{people}", produces = {"application/json"})
-    public ResponseEntity<?> getPopEqualToOrGreaterThan(@PathVariable int people){
+    public ResponseEntity<?> getPopEqualToOrGreaterThan(@PathVariable int people) {
         ArrayList<Country> rtnCountries = CountriesApplication.myCountryList.findCountries(c -> c.getPopulation() >= people);
         return new ResponseEntity<>(rtnCountries, HttpStatus.OK);
     }
-
+    // localhost:2019/population/min
     @GetMapping(value = "/min", produces = {"application/json"})
     public ResponseEntity<?> getCountryWithLeastPop() {
         ArrayList<Country> rtnCountries = CountriesApplication.myCountryList.countryList;
-        rtnCountries.sort((c1, c2) -> (int) c1.getPopulation() - c2.getPopulation());
-
+        rtnCountries.sort((c1, c2) -> c1.getPopulation() - c2.getPopulation());
         return new ResponseEntity<>(rtnCountries.get(0), HttpStatus.OK);
-
-
     }
+    // localhost:2019/population/max
     @GetMapping(value = "/max", produces = {"application/json"})
     public ResponseEntity<?> getCountryWithMostPop() {
         ArrayList<Country> rtnCountries = CountriesApplication.myCountryList.countryList;
-        rtnCountries.sort((c1, c2) ->  c2.getPopulation() - c1.getPopulation());
-
+        rtnCountries.sort((c1, c2) -> c2.getPopulation() - c1.getPopulation());
         return new ResponseEntity<>(rtnCountries.get(0), HttpStatus.OK);
     }
 }
